Working on a possible suspension layout.
The idea is to have 1 wheel size worth of travel. Having 1/2 of the travel compression and 1/2 of the travel in extension so that we keep good contact with the ground.
The mechanism should constrain the motion path to a vertical path so that steering with the suspension compressed doesn't lead to deviations in the odometry etc..
The current 'design' is based on the pantograph mechanism for electric trains because these move vertically only.